Support MCP Installation Deeplinks
Deeplinks in the form of `tome://install-mcp-server?payload=<payload>`
will open an installation page prompting the user to confirm (see
below).

## Payload Schema

The `payload` passed must be a url encoded JSON object. Example JSON
object:

```json
{
    "version": "1",
    "command": "uvx",
    "args": [
        "mcp-server-fetch"
    ],
    "env": {
        "FETCH_KEY": "123"
    }
}
```

All keys and values must be strings, as this ultimately becomes a CLI
command.

### Version

The `version` key is a future-proofing mechanism for if/when we make
backwards-incompatible changes to the schema (good call @camfulton).