Adds OpenAI support
Refactors the dispatch logic to be more engine-agnostic.

Engines are a backend to an LLM provider, like Ollama, OpenAI, etc. Each
Engine has a Client that is responsible for actually talking to the
backend.

Models now have an `id` in the format of `<engine>:<model-name>`. So for
example, `ollama:qwen3:14b`.

Models, in the model dropdown in chat, are now grouped and organized by
Engine. So there's an Ollama section, an OpenAI section, etc.