Convert Tome Messages into explicit Engine Messages
## Explicit Message Interfaces

Instead of trying to mash all Messages into a single, superset, shape
that fits all Engines – we'll use Message as the canonical shape and
explicitly convert to what Engines expect.

This means that each Engine will have an accompanying
`lib/models/messages/<engine>.ts` module, that exposes (at least) a
`from` function to convert from an `IMessage` to a message of the
Engine.

## Tool Call ID

The OpenAI API gives tool calls a unique identifier. The response
messages must reference that id via `tool_call_id`. This adds that to
messages AND fills it in where it would not normally (like from the
Ollama API, for example)