Handle missing Engines and Models
Now that we have multiple engines, the "Cannot connect to Ollama" no
longer makes sense, on it's own, so it was removed.

The problem is that it just explodes now when it can't connect to
_anything_.

This change handles startup when no models are present and shows an
appropriate error under the model selector in that scenario.