Resync models after Engine change
Resyncs all models when an Engine is updated. For example, when you
change the Ollama URL between two functioning Ollama instances, we need
to fully refetch the list of models it exposes.

This is done via an `afterUpdate` on Engine. That way, anytime an Engine
is changed in anyway, we re-fetch all models for all engines.

Also restores the functionality where new Sessions get the user's
configured "default" model.