Fixes Updater
Previously, we were pre-emptively showing the Restart button in the
Updater UI. We were (erroneously) doing this when the download was
complete, but before the installation process finished.

This was just caused by not fully understanding how the Tauri updater
plugin worked and consuming it incorrectly in TS. Go me.

This led to being able to restart the app before the new version was
installed, which led to the updater UI to pop up when the app restart.
It'd get stuck in this loop indefinitely.

This fixes it by waiting until the entire download + install process is
finished.