Fix missing model error
When a User doesn't have Ollama setup AND doesn't have a model
favorited, the app attempts to pull the first Engine's first model.

The problem is that the first Engine is ALWAYS Ollama. If the user
doesn't have Ollama setup, it will never resolve.

Instead, fallback to the first _Model_, regardless of which Engine is
belongs.