Scrollbars are dumb
The `overflow: scroll` was forcing the rendering of the scrollbar
gutter. Whether that gutter affects the rest of the elements or not is
system-set – not something controllable via CSS.

Because we had the `overflow: scroll` there was almost always extra
padding in `Lists` where the scrollbar _would_ go, but most of the time
there wasn't one because the list wasn't long enough.

This removes that space unless the List truly needs to scroll.