App/Relay -> McpServer UI overhaul
Managing associated McpServer records to App/Relay is inherently
complicated. We need to keep track of already persisted records, records
that need to be persisted on save, records that need to be deleted on
save, etc.

This change separates each of those categories out into their own
`$state()` object in the corresponding `Editor` components.

It starts with a list of "savable" McpServers. These are a combination
of persisted `McpServer` records and _copies_ of any "base" McpServers
not in the list of persisted ones.

A "base" McpServer is one managed in the MCP tab of the app. These are
never associated with an App or Relay. They for Chat, but are also the
default list of possible MCP servers for Apps/Relays.

We don't want to modify these records directly, hence the _copies_ of
them along side already associated records.